requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiohttp-cors==0.7.0
aiosignal==1.2.0
alembic==1.8.0
anyio==3.6.1
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
bentoml==1.0.0rc1
boto3==1.24.6
botocore==1.27.6
cattrs==22.1.0
Cerberus==1.3.4
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
circus==0.17.1
click==8.1.3
cloudpickle==2.1.0
colorama==0.4.4
commonmark==0.9.1
configparser==5.2.0
contextlib2==21.6.0
Cython==0.29.28
deepmerge==1.0.1
Deprecated==1.2.13
docker==5.0.3
exceptiongroup==1.0.0rc8
filelock==3.7.1
Flask==2.1.2
frozenlist==1.3.0
fs==2.4.16
funcy==1.17
future==0.18.2
gensim==4.2.0
grpcio==1.46.3
grpcio-tools==1.46.3
gunicorn==20.1.0
h11==0.13.0
humanfriendly==10.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
joblib==1.1.0
langdetect==1.0.9
Mako==1.2.0
MarkupSafe==2.1.1
multidict==6.0.2
nltk==3.7
numexpr==2.8.1
numpy==1.22.4
opentelemetry-api==1.9.0
opentelemetry-instrumentation==0.28b0
opentelemetry-instrumentation-aiohttp-client==0.28b0
opentelemetry-instrumentation-asgi==0.28b0
opentelemetry-sdk==1.9.0
opentelemetry-semantic-conventions==0.28b0
opentelemetry-util-http==0.28b0
packaging==21.3
pandas==1.4.2
pathspec==0.9.0
pep517==0.12.0
pip-tools==6.6.2
prometheus-client==0.13.1
protobuf==3.20.1
psutil==5.9.1
Pygments==2.12.0
pyLDAvis==3.3.1
pynvml==11.4.1
pyparsing==3.0.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-json-logger==2.0.2
python-multipart==0.0.5
pytz==2022.1
pywin32==227
PyYAML==6.0
pyzmq==23.1.0
regex==2022.6.2
requests==2.28.0
rich==12.4.4
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
schema==0.7.5
scikit-learn==1.1.1
scipy==1.8.1
simple-di==0.1.5
six==1.16.0
sklearn==0.0
smart-open==6.0.0
sniffio==1.2.0
SQLAlchemy==1.3.24
SQLAlchemy-Utils==0.36.5
starlette==0.20.3
tabulate==0.8.9
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.25.11
uvicorn==0.17.6
websocket-client==1.3.2
Werkzeug==2.1.2
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
